Changed top level element from mouseAnatomy to emapAnatomy.
Why?  I have embraced the term "EMAP" to describe the XML produced by this script, both in the names of the files that are produced, as well as in the names of the files that are producing it.  Renaming the top level element is consistent.

Why did I embrace EMAP?  I think it makes for better branding.